Add Docker configuration and update ProductController references
A Docker configuration file has been added to setup a workflow for building and pushing Docker images. The file also prepares the environment for setting up Docker buildx in the GitHub repository. The reference to the ProductController in the api.php route file has been updated to use the class directly.